Organize and document simulation.tensor
