+++ qualtran/simulation/tensor/__init__.py
@@ -1,0 +1,18 @@
+#  Copyright 2023 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+
+from ._dense import bloq_to_dense, get_right_and_left_inds
+from ._flattening import bloq_has_custom_tensors, flatten_for_tensor_contraction
+from ._quimb import cbloq_as_contracted_tensor, cbloq_to_quimb
