+++ qualtran/simulation/tensor/_quimb.py
@@ -11,15 +11,11 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-"""Functionality for the `Bloq.tensor_contract()` protocol."""
-
 import itertools
-from typing import Dict, List, Optional, Set, Tuple
+from typing import Dict, Optional, Set, Tuple
 
 import numpy as np
 import quimb.tensor as qtn
-from numpy.typing import NDArray
 
 from qualtran import (
     Bloq,
@@ -29,7 +25,6 @@ from qualtran import (
     DanglingT,
     LeftDangle,
     RightDangle,
-    Signature,
     Soquet,
     SoquetT,
 )
@@ -38,28 +33,6 @@ from qualtran._infra.composite_bloq import (
     _flatten_soquet_collection,
     _get_flat_dangling_soqs,
 )
-
-
-def bloq_has_custom_tensors(bloq: Bloq) -> bool:
-    """Whether this bloq declares custom tensors by overriding `.add_my_tensors(...)`.
-
-    This is a heuristic that checks that the method is overriden.
-    """
-    return not bloq.add_my_tensors.__qualname__.startswith(
-        'Bloq.'
-    ) and not bloq.add_my_tensors.__qualname__.startswith('GateWithRegisters.')
-
-
-def flatten_for_tensor_contraction(bloq: Bloq, max_depth: int = 1_000) -> CompositeBloq:
-    """Flatten a (composite) bloq as much as possible to enable efficient tensor contraction.
-
-    By default, bloqs without custom tensors will be contracted to a dense tensor using their
-    decomposition and then that dense tensor will be used in the enclosing tensor network.
-    To allow a more efficient contraction ordering, use this function to decompose-and-flatten
-    as much as possible before starting the tensor contraction.
-    """
-    cbloq = bloq.as_composite_bloq()
-    return cbloq.flatten(lambda binst: not bloq_has_custom_tensors(binst.bloq), max_depth=max_depth)
 
 
 def cbloq_to_quimb(
@@ -141,55 +114,9 @@ def cbloq_to_quimb(
     return tn, fix
 
 
-def get_right_and_left_inds(signature: Signature) -> List[List[Soquet]]:
-    """Return right and left indices.
-
-    In general, this will be returned as a list of length-2 corresponding
-    to the right and left indices, respectively. If there *are* no right
-    or left indices, that entry will be omitted from the returned list.
-
-    Right indices come first to match the quantum computing / matrix multiplication
-    convention where U_tot = U_n ... U_2 U_1.
-    """
-    inds = []
-    rsoqs = _get_flat_dangling_soqs(signature, right=True)
-    if rsoqs:
-        inds.append(rsoqs)
-    lsoqs = _get_flat_dangling_soqs(signature, right=False)
-    if lsoqs:
-        inds.append(lsoqs)
-    return inds
-
-
-def _cbloq_to_dense(cbloq: CompositeBloq) -> NDArray:
-    """Return a contracted, dense ndarray representing the composite bloq.
-
-    The public version of this function is available as the `CompositeBloq.tensor_contract()`
-    method.
-
-    This constructs a tensor network and then contracts it according to the cbloq's registers,
-    i.e. the dangling indices. The returned array will be 0-, 1- or 2- dimensional. If it is
-    a 2-dimensional matrix, we follow the quantum computing / matrix multiplication convention
-    of (right, left) indices.
-
-    For more fine grained control over the final shape of the tensor, use
-    `cbloq_to_quimb` and `TensorNetwork.to_dense` directly.
-    """
-    tn, _ = cbloq_to_quimb(cbloq)
-    inds = get_right_and_left_inds(cbloq.signature)
-
-    if inds:
-        return tn.to_dense(*inds)
-
-    return tn.contract()
-
-
-def _cbloq_as_contracted_tensor_data_and_inds(
-    cbloq: CompositeBloq,
-    signature: Signature,
-    incoming: Dict[str, SoquetT],
-    outgoing: Dict[str, SoquetT],
-) -> Tuple[NDArray, List[Soquet]]:
+def cbloq_as_contracted_tensor(
+    cbloq: CompositeBloq, incoming: Dict[str, SoquetT], outgoing: Dict[str, SoquetT], tags
+) -> qtn.Tensor:
     """`add_my_tensors` helper for contracting `cbloq` and adding it as a dense tensor.
 
     First, we turn the composite bloq into a TensorNetwork with `cbloq_to_quimb`. Then
@@ -200,6 +127,7 @@ def _cbloq_as_contracted_tensor_data_and_inds(
     # Turn into a dense ndarray, but instead of folding into a 1- or 2-
     # dimensional state/effect or unitary; we keep all the indices as
     # distinct dimensions.
+    signature = cbloq.signature
     rsoqs = _get_flat_dangling_soqs(signature, right=True)
     lsoqs = _get_flat_dangling_soqs(signature, right=False)
     inds_for_contract = rsoqs + lsoqs
@@ -215,5 +143,4 @@ def _cbloq_as_contracted_tensor_data_and_inds(
     inds_for_adding = _flatten_soquet_collection(itertools.chain(osoqs, isoqs))
     assert len(inds_for_adding) == len(inds_for_contract)
 
-    # Data and `inds_for_adding` can be used as `Tensor(data, inds)` arguments.
-    return data, inds_for_adding
+    return qtn.Tensor(data=data, inds=inds_for_adding, tags=tags)
